pyproject.toml
[project]
name = "tiny-engine"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11 <3.13"
dependencies = [
    "torch (>=2.7.1,<3.0.0)",
    "diffusers (>=0.33.1,<0.34.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "stable-baselines3[extra] (>=2.6.0,<3.0.0)",
    "transformers (>=4.52.4,<5.0.0)",
    "accelerate (>=1.7.0,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "vizdoom (>=1.2.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
